meta: Move embedded metadata format read functions into a module.
These formats are not standalone and are embedded into other
containers. Therefore, they do not implement the MetadataReader
trait and are instead provided as read and/or parse functions.
