+++ symphonia-metadata/src/embedded.rs
@@ -1,0 +1,12 @@
+// Symphonia
+// Copyright (c) 2019-2022 The Project Symphonia Developers.
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Reading and parsing of metadata formats that are embedded into other containers.
+
+pub mod flac;
+pub mod riff;
+pub mod vorbis;
