+++ symphonia-metadata/src/lib.rs
@@ -15,10 +15,8 @@
 #![allow(clippy::manual_range_contains)]
 
 pub mod ape;
-pub mod flac;
+pub mod embedded;
 pub mod id3v1;
 pub mod id3v2;
 pub mod itunes;
-pub mod riff;
 pub mod utils;
-pub mod vorbis;
